mnist/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
chex==0.1.90
comm==0.2.3
cramjam==2.11.0
debugpy==1.8.16
decorator==5.2.1
einops==0.8.1
etils==1.13.0
executing==2.2.0
fastparquet==2024.11.0
filelock==3.18.0
flax==0.11.1
fsspec==2025.7.0
hf-xet==1.1.7
huggingface-hub==0.34.4
humanize==4.12.3
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jax==0.7.0
jax-metrics==0.2.5
jaxlib==0.7.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.1
nest-asyncio==1.6.0
numpy==2.3.2
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.21
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.4
rich==14.1.0
scipy==1.16.1
simple-pytree==0.2.2
simplejson==3.20.1
six==1.17.0
stack-data==0.6.3
tensorstore==0.1.76
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.10
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
zipp==3.23.0
